Move createBasicEnvironment as test-util (#7560)
