+++ packages/astro/src/core/render/index.ts
@@ -2,6 +2,6 @@ export { createRenderContext } from './context.js';
 export type { RenderContext } from './context.js';
 export { renderPage } from './core.js';
 export type { Environment } from './environment';
-export { createBasicEnvironment, createEnvironment } from './environment.js';
+export { createEnvironment } from './environment.js';
 export { getParamsAndProps } from './params-and-props.js';
 export { loadRenderer, loadRenderers } from './renderer.js';
