+++ packages/astro/test/units/render/jsx.test.js
@@ -6,14 +6,9 @@ import {
 	renderSlot,
 } from '../../../dist/runtime/server/index.js';
 import { jsx } from '../../../dist/jsx-runtime/index.js';
-import {
-	createBasicEnvironment,
-	createRenderContext,
-	renderPage,
-	loadRenderer,
-} from '../../../dist/core/render/index.js';
+import { createRenderContext, renderPage, loadRenderer } from '../../../dist/core/render/index.js';
 import { createAstroJSXComponent, renderer as jsxRenderer } from '../../../dist/jsx/index.js';
-import { defaultLogging as logging } from '../../test-utils.js';
+import { createBasicEnvironment } from '../test-utils.js';
 
 const createAstroModule = (AstroComponent) => ({ default: AstroComponent });
 const loadJSXRenderer = () => loadRenderer(jsxRenderer, { import: (s) => import(s) });
@@ -23,7 +18,6 @@ describe('core/render', () => {
 		let env;
 		before(async () => {
 			env = createBasicEnvironment({
-				logging,
 				renderers: [await loadJSXRenderer()],
 			});
 		});
